Transpose data back again before returning to R
Even though the data is passed by reference, it is still modified
in-place.

See https://stackoverflow.com/questions/24112893/within-c-functions-how-are-rcpp-objects-passed-to-other-functions-by-referen).

So when the data is transposed for ComputeMST, it must be transposed
back again before returning to R, otherwise the x,y coordinates will be
jumbled in the data frame returned to the user.